requirements.txt
accelerate==1.1.1
adapter-transformers==4.0.0
adapters==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-common==1.1.28
backoff==2.2.1
blessed==1.20.0
blinker==1.9.0
boilerpy3==1.0.7
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
construct==2.5.3
contourpy==1.3.1
cwcwidth==0.1.9
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.2
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==5.1.3
django-haystack==3.3.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
enlighten==1.10.1
Events==0.5
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.7.2
farm-haystack==1.26.4
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
graphql-core==3.2.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
haystack==0.42
haystack-ai==2.7.0
haystack-experimental==0.3.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
inflect==7.4.0
ipython==8.28.0
isodate==0.7.2
itsdangerous==2.2.0
jarowinkler==1.2.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langchain-text-splitters==0.3.4
langchainplus-sdk==0.0.20
langdetect==1.0.9
langsmith==0.0.83
latexcodec==3.0.0
lazy-imports==0.3.1
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.1
marshmallow==3.23.1
matplotlib-inline==0.1.7
mecab-python3==1.0.9
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
mtdata==0.4.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.55.3
orjson==3.10.12
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pefile==2024.8.26
peft==0.11.2.dev0
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
portalocker==2.3.0
posthog==3.7.4
prefixed==0.9.0
prompt_toolkit==3.0.48
prompthub-py==4.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybtex==0.24.0
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.27.1
pydantic-settings==2.6.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-frontmatter==1.1.0
python-magic==0.4.27
python-pptx==1.0.0
python-ptrace==0.9.9
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
rapidfuzz==2.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-cache==0.9.8
requests-toolbelt==1.0.0
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.0.0
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
sklearn==0.0.post12
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.3.0+cu118
torchao==0.6.1
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
translate-toolkit==3.14.3
trio==0.27.0
triton==2.3.0
typeguard==4.4.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unidic-lite==1.0.8
url-normalize==1.4.3
urllib3==1.26.20
validators==0.34.0
wandb==0.18.7
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.0
zipp==3.21.0
